tweak: support resizing TTY
